# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Weather API",
      "description": "API for accessing weather data",
      "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
      "/api/weather": {
        "get": {
          "summary": "Get weather data",
          "parameters": [
            {
              "name": "station_id",
              "in": "query",
              "type": "string",
              "required": false
            },
            {
              "name": "date",
              "in": "query",
              "type": "string",
              "required": false
            },
            {
              "name": "page",
              "in": "query",
              "type": "integer",
              "required": false,
              "default": 1
            },
            {
              "name": "per_page",
              "in": "query",
              "type": "integer",
              "required": false,
              "default": 10
            }
          ],
          "responses": {
            "200": {
              "description": "A list of weather data",
              "schema": {
                "type": "object",
                "properties": {
                  "total": { "type": "integer" },
                  "page": { "type": "integer" },
                  "per_page": { "type": "integer" },
                  "data": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "station_id": { "type": "string" },
                        "date": { "type": "string" },
                        "max_temp": { "type": "number" },
                        "min_temp": { "type": "number" },
                        "precipitation": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/weather/stats": {
        "get": {
          "summary": "Get weather statistics",
          "parameters": [
            {
              "name": "station_id",
              "in": "query",
              "type": "string",
              "required": false
            },
            
            {
              "name": "page",
              "in": "query",
              "type": "integer",
              "required": false,
              "default": 1
            },
            {
              "name": "per_page",
              "in": "query",
              "type": "integer",
              "required": false,
              "default": 10
            }
          ],
          "responses": {
          "200": {
            "description": "Weather statistics",
            "schema": {
              "type": "object",
              "properties": {
                "total": { "type": "integer" },
                "page": { "type": "integer" },
                "per_page": { "type": "integer" },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "station_id": { "type": "string" },
                      "year": { "type": "integer" },
                      "avg_max_temp": { "type": "number" },
                      "avg_min_temp": { "type": "number" },
                      "total_precipitation": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
  